replace StaticPrecompiler with django-pipline.
Django pipline can concatinate and uglyfi the css/javascript files, and it is possible to configure it so it do not neet ruby-sass with libsasscompiler. if we get rid of coffee we can run again without external compilers. This is work in Progress and only works in debug mode for now