Add vote choice filter to poll overview
This feature is strongly based on the filter feature from opatut/dudel.

This commit allows a user to filter the vote choices (columns) in the
poll overview based on the column score. The user can filter the poll
based on a threshold percentage or to show only the best x columns.